Updates process for syncing login state between web instances
- Updated "login sync" process so that the "selecting to browse as guest", "logging in" and "logging out" processes cause the "login state" to be synced between web-based app instances., with the refreshing of all displayed Feeds being part of the process.
- Created a new `FeedList.RefreshAllFeeds()` method for refreshing all Feeds displayed in the `FeedColumn` component.
- Fixed issue where the currently logged in User's avatar would not be loaded in the `UserButton` component after logging in.
